requirements.txt
flask==3.1.1
openai==1.3.7
python-dotenv==1.0.1
markdown==3.6
requests==2.32.3
pydub==0.25.1
